+++ common/src/main/java/dev/terminalmc/flashside/config/Config.java
@@ -43,6 +43,12 @@ public class Config {
     }
 
     public static class Options {
+        public static final boolean leftSideDefault = false;
+        public boolean leftSide = true;
+        
+        public static final boolean editTitleScreenDefault = false;
+        public boolean editTitleScreen = true;
+        
         public static final int startRowDefault = 1;
         public int startRow = startRowDefault;