+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3.0
+
+- Added option to place buttons on left side
+- Added option to move Flashback title screen button to left side
+
 ## 0.2.0
 
 - Redesigned button injection using MixinSquared